+++ example/src/main/java/com/mercadopago/android/px/di/preference/InitializationDataPreferences.kt
@@ -1,7 +1,9 @@
 package com.mercadopago.android.px.di.preference
 
 import android.content.SharedPreferences
+import com.mercadopago.android.px.feature.custom_initialize.*
 import com.mercadopago.android.px.feature.custom_initialize.ConfigurationBooleanData
+import com.mercadopago.android.px.feature.custom_initialize.ConfigurationDataType
 import com.mercadopago.android.px.feature.custom_initialize.ConfigurationStringData
 import com.mercadopago.android.px.feature.custom_initialize.InitializationData
 import com.mercadopago.android.px.internal.util.TextUtil
@@ -15,6 +17,7 @@ internal class InitializationDataPreferences(private val preferences: SharedPref
             putString(PREFERENCE_ID, data.preferenceId.value)
             putString(ACCESS_TOKEN, data.accessToken.value)
             putBoolean(ONE_TAP, data.oneTap.value)
+            putString(PROCESSOR_TYPE, data.processor.value.name)
             apply()
         }
     }
@@ -24,7 +27,9 @@ internal class InitializationDataPreferences(private val preferences: SharedPref
         ConfigurationStringData.PublicKey(preferences.getString(PUBLIC_KEY, TextUtil.EMPTY) ?: TextUtil.EMPTY),
         ConfigurationStringData.PreferenceId(preferences.getString(PREFERENCE_ID, TextUtil.EMPTY) ?: TextUtil.EMPTY),
         ConfigurationStringData.AccessToken(preferences.getString(ACCESS_TOKEN, TextUtil.EMPTY) ?: TextUtil.EMPTY),
-        ConfigurationBooleanData.OneTap(preferences.getBoolean(ONE_TAP, true)))
+        ConfigurationBooleanData.OneTap(preferences.getBoolean(ONE_TAP, true)),
+        ConfigurationDataType.Processor(ProcessorType.valueOf(
+            preferences.getString(PROCESSOR_TYPE, ProcessorType.DEFAULT.name) ?: ProcessorType.DEFAULT.name)))
 
     companion object {
         private const val DEFAULT_LOCALE = "en-US"
@@ -33,5 +38,6 @@ internal class InitializationDataPreferences(private val preferences: SharedPref
         private const val PREFERENCE_ID = "preference_id"
         private const val ACCESS_TOKEN = "access_token"
         private const val ONE_TAP = "one_tap"
+        private const val PROCESSOR_TYPE = "processor_type"
     }
 }