+++ px-addons/src/main/java/com/mercadopago/android/px/addons/internal/TrackingDefaultBehaviour.java
@@ -2,6 +2,7 @@ package com.mercadopago.android.px.addons.internal;
 
 import android.support.annotation.NonNull;
 import com.mercadopago.android.px.addons.TrackingBehaviour;
+import com.mercadopago.android.px.addons.model.Track;
 import java.util.Map;
 
 public class TrackingDefaultBehaviour implements TrackingBehaviour {
@@ -19,5 +20,10 @@ public class TrackingDefaultBehaviour implements TrackingBehaviour {
     @Override
     public void onEvent(@NonNull final String path, @NonNull final Map<String, ?> data) {
         //Do nothing
+    }
+
+    @Override
+    public void track(@NonNull final Track track) {
+        //do nothing
     }
 }