+++ px-addons/src/main/java/com/mercadopago/android/px/addons/model/internal/AvailableFeature.kt
@@ -1,0 +1,5 @@
+package com.mercadopago.android.px.addons.model.internal
+
+import java.io.Serializable
+
+data class AvailableFeature(val id: String, val enabled: Boolean) : Serializable