+++ px-addons/src/main/java/com/mercadopago/android/px/addons/model/internal/Variant.kt
@@ -1,0 +1,5 @@
+package com.mercadopago.android.px.addons.model.internal
+
+import java.io.Serializable
+
+data class Variant(val id: Int, val name: String, val availableFeatures: List<AvailableFeature>?) : Serializable