+++ px-services/src/main/java/com/mercadopago/android/px/model/CardInformation.kt
@@ -1,0 +1,16 @@
+package com.mercadopago.android.px.model
+
+import java.io.Serializable
+
+interface CardInformation : Serializable {
+    val expirationMonth: Int?
+    val expirationYear: Int?
+    val cardHolder: Cardholder?
+    var lastFourDigits: String?
+    var firstSixDigits: String?
+    fun getSecurityCodeLength(): Int?
+
+    companion object {
+        const val CARD_NUMBER_MAX_LENGTH = 16
+    }
+}